refactor: edit gatsby-node.js
